Upgrade to 64-bit, yeah! arrays.length now returns ulongs, which need explicit conversion to ints. Can't stringify reals. json.Value now has opIndexAssign (fixes bug, why did it work before..?). Removed dependency on opencl, cl4d and the win32-binding. Upgraded to dmd 2.062, notable fixes include angrier shadow declarations and ref parameters. Also algorithm.remove(alias pred) resolves to the wrong remove, so remade code to use the other remove. Removed raycastgpu (Can reimplement with compute shaders easily). Added windows.d which replaces the win32-binding.

+++ cgy/modelparser/cgyparser.d
@@ -30,7 +30,7 @@ final class Vert {
 }
 
 final class Tri {
-    size_t[3] verts;
+    uint[3] verts;
 }
